nifty_trading_agent/requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.3.0
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.13.0
yfinance>=0.2.0
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
PyYAML>=6.0
python-dateutil>=2.8.0
tqdm>=4.64.0
joblib>=1.2.0
openpyxl>=3.1.0
xgboost>=1.7.0
nltk>=3.8.0
textblob>=0.17.0
vaderSentiment>=3.3.0
psycopg2-binary>=2.9.0
duckdb>=0.9.0
imbalanced-learn>=0.11.0
